UPSTREAM: fuse: allow fine grained attr cache invaldation
This patch adds the infrastructure for more fine grained attribute
invalidation.  Currently only 'atime' is invalidated separately.

The use of this infrastructure is extended to the statx(2) interface, which
for now means that if only 'atime' is invalid and STATX_ATIME is not
specified in the mask argument, then no GETATTR request will be generated.

Change-Id: I74f39e88df73d3bd3fa10368de448db1bb9cae96
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: rajdeep-3305 <rajdeepbiswas3305@gmail.com>