UPSTREAM: fuse: switch to ->free_inode()
fuse_destroy_inode() is gone - sanity checks that need the stack
trace of the caller get moved into ->evict_inode(), the rest joins
the RCU-delayed part which becomes ->free_inode().

While we are at it, don't just pass the address of what happens
to be the first member of structure to kmem_cache_free() -
get_fuse_inode() is there for purpose and it gives the proper
container_of() use.  No behaviour change, but verifying correctness
is easier that way.

Change-Id: I64332cedbc93df162e34851edb268f5703dd730a
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: rajdeep-3305 <rajdeepbiswas3305@gmail.com>