[ALPS08181638] GPU: GPUCORE-34917 for CVE-2023-28469
[Description]
After the GPU page table update when pages are mapped/unmapped, MMU
command is sent for the L2 cache maintenance and MMU TLB invalidation.
The affected GPU virtual range is specified in the LOCKADDR register
before the command is sent. Due to the alignment constraints, the lock
range specified can be more than the virtual range that was updated.

There was a bug in the calculation of locked range when the virtual
range straddled the 4GB boundary. The fls() function was used that
accepts a 32 bit value, but when the virtual range straddled 4GB the
value greater than 32 bit was passed. As a result the size of locked
range was calculated incorrectly.

This commit fixes the bug by using fls64() that takes an argument of
u64 type.

[Patch]
GPUCORE-34917: Fix lock region size calculation when virtual range
straddles 4GB

[ARM Ticket]
03422243

MTK-Commit-Id: c1c4a0e958bf7b44b2a87dcd5604547fdf29f42b

Signed-off-by: Daniel Tao <daniel.tao@mediatek.com>
CR-Id: ALPS08181638
Feature: ARM GPU DDK driver
Change-Id: If7ca2f43aed9c6a64d974bdfb89fc0ec419003e2
(cherry-pick from 77e65557b70ac6a4d193d4de7824fc1e5834b806)